renamed svm-js directory to only svm
