+++ src/bp/ml/svm/normalize-input.ts
